Crate atsamd21j

Peripheral access API for ATSAMD21J microcontrollers (generated using svd2rust v0.20.0 ( ))

You can find an overview of the generated API here.

API features to be included in the next svd2rust release can be generated by cloning the svd2rust repository, checking out the above commit, and running cargo doc --open.

Re-exports

pub use pac0 as pac1;

pub use pac0 as pac2;

pub use port as port_iobus;

pub use sercom0 as sercom1;

pub use sercom0 as sercom2;

pub use sercom0 as sercom3;

pub use sercom0 as sercom4;

pub use sercom0 as sercom5;

pub use tc3 as tc4;

pub use tc3 as tc5;

pub use tc3 as tc6;

pub use tc3 as tc7;

pub use tcc0 as tcc1;

pub use tcc0 as tcc2;

Modules

ac

Analog Comparators

adc

Analog Digital Converter

dac

Digital Analog Converter

dmac

Direct Memory Access Controller

dsu

Device Service Unit

eic

External Interrupt Controller

evsys

Event System Interface

gclk

Generic Clock Generator

generic

Common register and bit access and modify traits

hmatrix

HSB Matrix

i2s

Inter-IC Sound Interface

mtb

Cortex-M0+ Micro-Trace Buffer

nvmctrl

Non-Volatile Memory Controller

pac0

Peripheral Access Controller 0

pm

Power Manager

port

Port Module

rtc

Real-Time Counter

sercom0

Serial Communication Interface 0

sysctrl

System Control

tc3

Basic Timer Counter 3

tcc0

Timer Counter Control 0

usb

Universal Serial Bus

wdt

Watchdog Timer

Structs

AC

Analog Comparators

ADC

Analog Digital Converter

CBP

Cache and branch predictor maintenance operations

CPUID

CPUID

CorePeripherals

Core peripherals

DAC

Digital Analog Converter

DCB

Debug Control Block

DMAC

Direct Memory Access Controller

DSU

Device Service Unit

DWT

Data Watchpoint and Trace unit

EIC

External Interrupt Controller

EVSYS

Event System Interface

FPB

Flash Patch and Breakpoint unit

GCLK

Generic Clock Generator

HMATRIX

HSB Matrix

I2S

Inter-IC Sound Interface

ITM

Instrumentation Trace Macrocell

MPU

Memory Protection Unit

MTB

Cortex-M0+ Micro-Trace Buffer

NVIC

Nested Vector Interrupt Controller

NVMCTRL

Non-Volatile Memory Controller

PAC0

Peripheral Access Controller 0

PAC1

Peripheral Access Controller 1

PAC2

Peripheral Access Controller 2

PM

Power Manager

PORT

Port Module

PORT_IOBUS

Port Module (IOBUS)

Peripherals

All the peripherals

RTC

Real-Time Counter

SCB

System Control Block

SERCOM0

Serial Communication Interface 0

SERCOM1

Serial Communication Interface 1

SERCOM2

Serial Communication Interface 2

SERCOM3

Serial Communication Interface 3

SERCOM4

Serial Communication Interface 4

SERCOM5

Serial Communication Interface 5

SYSCTRL

System Control

SYST

SysTick: System Timer

TC3

Basic Timer Counter 3

TC4

Basic Timer Counter 4

TC5

Basic Timer Counter 5

TC6

Basic Timer Counter 6

TC7

Basic Timer Counter 7

TCC0

Timer Counter Control 0

TCC1

Timer Counter Control 1

TCC2

Timer Counter Control 2

TPIU

Trace Port Interface Unit

USB

Universal Serial Bus

WDT

Watchdog Timer

Enums

Interrupt

Enumeration of all the interrupts.

Constants

NVIC_PRIO_BITS

Number available in the NVIC for configuring priority