Type Definition atsamd21j::dmac::chctrlb::TRIGSRC_R

type TRIGSRC_R = R<u8, TRIGSRC_A>;

Reader of field TRIGSRC

Implementations

impl TRIGSRC_R

pub fn variant(&self) -> Variant<u8, TRIGSRC_A>

Get enumerated values variant

pub fn is_disable(&self) -> bool

Checks if the value of the field is DISABLE

pub fn is_sercom0_rx(&self) -> bool

Checks if the value of the field is SERCOM0_RX

pub fn is_sercom0_tx(&self) -> bool

Checks if the value of the field is SERCOM0_TX

pub fn is_sercom1_rx(&self) -> bool

Checks if the value of the field is SERCOM1_RX

pub fn is_sercom1_tx(&self) -> bool

Checks if the value of the field is SERCOM1_TX

pub fn is_sercom2_rx(&self) -> bool

Checks if the value of the field is SERCOM2_RX

pub fn is_sercom2_tx(&self) -> bool

Checks if the value of the field is SERCOM2_TX

pub fn is_sercom3_rx(&self) -> bool

Checks if the value of the field is SERCOM3_RX

pub fn is_sercom3_tx(&self) -> bool

Checks if the value of the field is SERCOM3_TX

pub fn is_sercom4_rx(&self) -> bool

Checks if the value of the field is SERCOM4_RX

pub fn is_sercom4_tx(&self) -> bool

Checks if the value of the field is SERCOM4_TX

pub fn is_sercom5_rx(&self) -> bool

Checks if the value of the field is SERCOM5_RX

pub fn is_sercom5_tx(&self) -> bool

Checks if the value of the field is SERCOM5_TX

pub fn is_tcc0_ovf(&self) -> bool

Checks if the value of the field is TCC0_OVF

pub fn is_tcc0_mc0(&self) -> bool

Checks if the value of the field is TCC0_MC0

pub fn is_tcc0_mc1(&self) -> bool

Checks if the value of the field is TCC0_MC1

pub fn is_tcc0_mc2(&self) -> bool

Checks if the value of the field is TCC0_MC2

pub fn is_tcc0_mc3(&self) -> bool

Checks if the value of the field is TCC0_MC3

pub fn is_tcc1_ovf(&self) -> bool

Checks if the value of the field is TCC1_OVF

pub fn is_tcc1_mc0(&self) -> bool

Checks if the value of the field is TCC1_MC0

pub fn is_tcc1_mc1(&self) -> bool

Checks if the value of the field is TCC1_MC1

pub fn is_tcc2_ovf(&self) -> bool

Checks if the value of the field is TCC2_OVF

pub fn is_tcc2_mc0(&self) -> bool

Checks if the value of the field is TCC2_MC0

pub fn is_tcc2_mc1(&self) -> bool

Checks if the value of the field is TCC2_MC1

pub fn is_tc3_ovf(&self) -> bool

Checks if the value of the field is TC3_OVF

pub fn is_tc3_mc0(&self) -> bool

Checks if the value of the field is TC3_MC0

pub fn is_tc3_mc1(&self) -> bool

Checks if the value of the field is TC3_MC1

pub fn is_tc4_ovf(&self) -> bool

Checks if the value of the field is TC4_OVF

pub fn is_tc4_mc0(&self) -> bool

Checks if the value of the field is TC4_MC0

pub fn is_tc4_mc1(&self) -> bool

Checks if the value of the field is TC4_MC1

pub fn is_tc5_ovf(&self) -> bool

Checks if the value of the field is TC5_OVF

pub fn is_tc5_mc0(&self) -> bool

Checks if the value of the field is TC5_MC0

pub fn is_tc5_mc1(&self) -> bool

Checks if the value of the field is TC5_MC1

pub fn is_tc6_ovf(&self) -> bool

Checks if the value of the field is TC6_OVF

pub fn is_tc6_mc0(&self) -> bool

Checks if the value of the field is TC6_MC0

pub fn is_tc6_mc1(&self) -> bool

Checks if the value of the field is TC6_MC1

pub fn is_tc7_ovf(&self) -> bool

Checks if the value of the field is TC7_OVF

pub fn is_tc7_mc0(&self) -> bool

Checks if the value of the field is TC7_MC0

pub fn is_tc7_mc1(&self) -> bool

Checks if the value of the field is TC7_MC1

pub fn is_adc_resrdy(&self) -> bool

Checks if the value of the field is ADC_RESRDY

pub fn is_dac_empty(&self) -> bool

Checks if the value of the field is DAC_EMPTY

pub fn is_i2s_rx_0(&self) -> bool

Checks if the value of the field is I2S_RX_0

pub fn is_i2s_rx_1(&self) -> bool

Checks if the value of the field is I2S_RX_1

pub fn is_i2s_tx_0(&self) -> bool

Checks if the value of the field is I2S_TX_0

pub fn is_i2s_tx_1(&self) -> bool

Checks if the value of the field is I2S_TX_1

pub fn is_tcc3_ovf(&self) -> bool

Checks if the value of the field is TCC3_OVF

pub fn is_tcc3_mc0(&self) -> bool

Checks if the value of the field is TCC3_MC0

pub fn is_tcc3_mc1(&self) -> bool

Checks if the value of the field is TCC3_MC1

pub fn is_tcc3_mc2(&self) -> bool

Checks if the value of the field is TCC3_MC2

pub fn is_tcc3_mc3(&self) -> bool

Checks if the value of the field is TCC3_MC3