Type Definition atsamd21j::gclk::clkctrl::GEN_R

type GEN_R = R<u8, GEN_A>;

Reader of field GEN

Implementations

impl GEN_R

pub fn variant(&self) -> Variant<u8, GEN_A>

Get enumerated values variant

pub fn is_gclk0(&self) -> bool

Checks if the value of the field is GCLK0

pub fn is_gclk1(&self) -> bool

Checks if the value of the field is GCLK1

pub fn is_gclk2(&self) -> bool

Checks if the value of the field is GCLK2

pub fn is_gclk3(&self) -> bool

Checks if the value of the field is GCLK3

pub fn is_gclk4(&self) -> bool

Checks if the value of the field is GCLK4

pub fn is_gclk5(&self) -> bool

Checks if the value of the field is GCLK5

pub fn is_gclk6(&self) -> bool

Checks if the value of the field is GCLK6

pub fn is_gclk7(&self) -> bool

Checks if the value of the field is GCLK7

pub fn is_gclk8(&self) -> bool

Checks if the value of the field is GCLK8