atsamd21j::nvmctrl::ctrlb

Type Alias ReadmodeR

pub type ReadmodeR = FieldReader<Readmodeselect>;

Field READMODE reader - NVMCTRL Read Mode

Aliased Type

struct ReadmodeR { /* private fields */ }

Implementations

impl ReadmodeR

pub const fn variant(&self) -> Option<Readmodeselect>

Get enumerated values variant

pub fn is_no_miss_penalty(&self) -> bool

The NVM Controller (cache system) does not insert wait states on a cache miss. Gives the best system performance.

pub fn is_low_power(&self) -> bool

Reduces power consumption of the cache system, but inserts a wait state each time there is a cache miss. This mode may not be relevant if CPU performance is required, as the application will be stalled and may lead to increase run time.

pub fn is_deterministic(&self) -> bool

The cache system ensures that a cache hit or miss takes the same amount of time, determined by the number of programmed flash wait states. This mode can be used for real-time applications that require deterministic execution timings.