Type Definition atsamd21j::pm::apbasel::APBADIV_R

type APBADIV_R = R<u8, APBADIV_A>;

Reader of field APBADIV

Implementations

impl APBADIV_R

pub fn variant(&self) -> APBADIV_A

Get enumerated values variant

pub fn is_div1(&self) -> bool

Checks if the value of the field is DIV1

pub fn is_div2(&self) -> bool

Checks if the value of the field is DIV2

pub fn is_div4(&self) -> bool

Checks if the value of the field is DIV4

pub fn is_div8(&self) -> bool

Checks if the value of the field is DIV8

pub fn is_div16(&self) -> bool

Checks if the value of the field is DIV16

pub fn is_div32(&self) -> bool

Checks if the value of the field is DIV32

pub fn is_div64(&self) -> bool

Checks if the value of the field is DIV64

pub fn is_div128(&self) -> bool

Checks if the value of the field is DIV128