atsamd21j::tcc0::ctrlbclr

Type Alias IdxcmdR

pub type IdxcmdR = FieldReader<Idxcmdselect>;

Field IDXCMD reader - Ramp Index Command

Aliased Type

struct IdxcmdR { /* private fields */ }

Implementations

impl IdxcmdR

pub const fn variant(&self) -> Idxcmdselect

Get enumerated values variant

pub fn is_disable(&self) -> bool

Command disabled: Index toggles between cycles A and B

pub fn is_set(&self) -> bool

Set index: cycle B will be forced in the next cycle

pub fn is_clear(&self) -> bool

Clear index: cycle A will be forced in the next cycle

pub fn is_hold(&self) -> bool

Hold index: the next cycle will be the same as the current cycle