Type Definition atsamd21j::tcc0::waveb::WAVEGENB_R

type WAVEGENB_R = R<u8, WAVEGENB_A>;

Reader of field WAVEGENB

Implementations

impl WAVEGENB_R

pub fn variant(&self) -> Variant<u8, WAVEGENB_A>

Get enumerated values variant

pub fn is_nfrq(&self) -> bool

Checks if the value of the field is NFRQ

pub fn is_mfrq(&self) -> bool

Checks if the value of the field is MFRQ

pub fn is_npwm(&self) -> bool

Checks if the value of the field is NPWM

pub fn is_dscritical(&self) -> bool

Checks if the value of the field is DSCRITICAL

pub fn is_dsbottom(&self) -> bool

Checks if the value of the field is DSBOTTOM

pub fn is_dsboth(&self) -> bool

Checks if the value of the field is DSBOTH

pub fn is_dstop(&self) -> bool

Checks if the value of the field is DSTOP