Type Definition atsamd21j::wdt::config::WINDOW_R

type WINDOW_R = R<u8, WINDOW_A>;

Reader of field WINDOW

Implementations

impl WINDOW_R

pub fn variant(&self) -> Variant<u8, WINDOW_A>

Get enumerated values variant

pub fn is_8(&self) -> bool

Checks if the value of the field is _8

pub fn is_16(&self) -> bool

Checks if the value of the field is _16

pub fn is_32(&self) -> bool

Checks if the value of the field is _32

pub fn is_64(&self) -> bool

Checks if the value of the field is _64

pub fn is_128(&self) -> bool

Checks if the value of the field is _128

pub fn is_256(&self) -> bool

Checks if the value of the field is _256

pub fn is_512(&self) -> bool

Checks if the value of the field is _512

pub fn is_1k(&self) -> bool

Checks if the value of the field is _1K

pub fn is_2k(&self) -> bool

Checks if the value of the field is _2K

pub fn is_4k(&self) -> bool

Checks if the value of the field is _4K

pub fn is_8k(&self) -> bool

Checks if the value of the field is _8K

pub fn is_16k(&self) -> bool

Checks if the value of the field is _16K